Muda estrutura de pastas e arquivos
