+++ lib/domain/models/cart_item_model.dart
