--- lib/providers/cupons_provider.dart
@@ -1,19 +1,0 @@
-import 'package:cafe_pra_ja/models/cupons_item_model.dart';
-import 'package:cafe_pra_ja/services/database_service.dart';
-import 'package:flutter/foundation.dart';
-
-class CuponsProvider with ChangeNotifier {
-  late final DatabaseService _dbService;
-
-  CuponsProvider() : _dbService = DatabaseService();
-
-  Stream<List<CuponsItemModel>> get cuponsStream {
-    return _dbService.getTodosOsCuponsStream().map((querySnapshot) {
-      final List<CuponsItemModel> listaDeCupons =
-          querySnapshot.docs.map((doc) {
-            return CuponsItemModel.fromFirestore(doc);
-          }).toList();
-      return listaDeCupons;
-    });
-  }
-}
